Upload of avvio.c and padre_treni.c
avvio.c check the arguments and print if you are in ETCS1 or ETCS2 (also if you are using MAPPA1 or MAPPA2).
padre_treni.c create 16 files, one for each piece of rail. It also create a directory where those files are placed. All files are initialized with a zero.