Created the Server part in RBC
The communication between trains and RBCServer work and they send each other all the stuffs but it seems that they don't communicate anymore after the first one. I think that there are some errors in the manage of the MAs and Stations data arrays. You can try to launch the two shell and you can see that some cells of the Stations' array are set at -3 and I don't know how to fix it